move ProfilesSelector into its own file
